improving organization of helper functions used to retrieve the server type

rh-pre-commit.version: 2.3.2
rh-pre-commit.check-secrets: ENABLED
